refactor(request-client): put related types in one file
