+++ src/types/index.ts
@@ -6,6 +6,7 @@ export * from './Job'
 export * from './JobDefinition'
 export * from './JobResult'
 export * from './Link'
+export * from './Login'
 export * from './SASjsConfig'
 export * from './RequestClient'
 export * from './Session'
@@ -13,3 +14,4 @@ export * from './UploadFile'
 export * from './PollOptions'
 export * from './WriteStream'
 export * from './ExecuteScript'
+export * from './errors'
